Benchmarks for dynamic variable access
Add some primitive benchmarks for variable access. I came to like the
criterion microbenchmarking framework so let's use it. Since the crate
does not have much API surface to cover, we measure the overhead for
accessing a dynamic variable for reading and writing. Difference between
the benchmarks for static and dynamic is the overhead spent on thread-
local storage and all that's in-between.

The overhead turns out to be relatively small: around 5 nanoseconds on
my machine. However, it is also quite large in the perspective: static
variable access takes about 0.5 nanoseconds, therefore dynamic variables
introduce about x10 slowdown.