Separate build and run steps
Well... Benchmarking requires criterion crate which pull in quite
a few additional development dependencies. Since benchmarks are
using the same --test configuration as tests, all this stuff gets
built even when we run "cargo test" and that takes a lot of time.

Let's split the building and running parts of tests so that the CI
report looks more useful and accurately reflects time spend on
running our tests, as opposed to building other people's code.