Correct usage of black_box()
We use black_box() to ensure that the compiler does not optimize out
assignments to COUNTER in the static variable tests.

Clippy did not like that we pass a unit value into the function,
since that still allows the compiler to perform optimizations.

Let's not fool around with that, using black_box() as a barrier
to ensure that the compiler emits both assignments to COUNTER.