More accurate benchmarks
Make sure that the compiler actually emits memory loads and stores
in attempt to accurately measure the overhead of accessing dynamic
variables. We'd like to make sure that the compiler outputs the
same code in both cases (barring static/dynamic differences),
and does not try to inline anything unnecessary or cache values
in registers and optimize memory loads and stores.