More strict value setting
Replace the trait bound on the set() method to be just reference instead
of using Borrow<T> which allows either references or owned values. This
is not really a security issue, but it has some weird readability issues
with types that implement Copy (and those are most likely candidates for
being used as dynamic variables). Using Borrow<T> allows the following:

    #[test]
    fn apparent_aliasing_violation() {
        fluid_let!(static VARIABLE: i32);

        let mut value = 0;
        VARIABLE.set(value, || {
            VARIABLE.get(|value_2| {
                VARIABLE.set(value_2.unwrap_or(&0), || {
                    VARIABLE.get(|value_3| {
                        value = 10;
                        assert_eq!(value_3, Some(&10));
                        // actually assetion fails, value_3 == 0
                    })
                });
            });
        });
    }

which *looks like* a violation of Rust aliasing rules, but actually it
isn't. It seems that you can assign new value to the original "value"
while still having references to it inside get() calls, but that's not
what's happening here. You actually assign to an implicit temporary copy
of the "value" made in the first set() call.

Let's deny this and require the user to explicitly pass a reference.
This way we don't introduce any ambiguity and the compiler will warn
about actual aliasing violations where you do try to assign to the
variable you use for dynamic value.

While we're here, make miscellaneous updates to the documentation
to make it clear what guarantees we provide and how to use dynamic
variables. And clean up terminology a bit.