Convenience copying accessors
Often dynamic variables are used for simple values like boolean flags
or integer values. In this case we do not really need to use the weird
closure-based accessor mechanism and are totally fine with a quick copy.

Provide copied() and cloned() accessors similar to Option<T> methods
which copy and clone the inner Option value.