Convenience setter macro
It turns out that Rust macros are actually splicing so we are able to
write a macro that make scoped assignments easier. Normally, this is
what users would probably do.

Introduce a fluid_set! macro which can reduce the amount of nesting
and acts as a scoped RAII-style guard for binding dynamic variables.

We cannot expose DynamicCellGuard to the public because it is tricky
to use safely. You absolutely must not drop your guard and drop the
guard before you drop the inner guard. Otherwise it leads to bad
behavior described in the cell_unsafe_set_get_usage() test. Exposing
the guard to the user directly technically allows them to do anything,
like calling drop() for them, and this breaks memory safety. Mutex
guards from the standard library get away with it because you are
not allowed to lock the same mutex twice from the same thread. But
that's obviously not the case for dynamic variables.

Hiding the guard under the macro resolves this issue since the user
does not have direct access to it and thus is unable to drop it.
Note, however, that we still have to make relevant APIs public
because that's how Rust macros work. Though, normally the user
should not be able to access them. And if they do then they are
marked with "unsafe".