Miscellaneous cleanup (#10)
* Move tests into the crate source

We don't really need integration tests that much. The API is documented
in rustdocs, so we can test all the stuff we need right in the crate
source code. Integration tests usually serve as example code, but ours
are not really that clean. Just read the API docs.

* Avoid warnings about guard being unused

Now rustc is able to see that the guard is not really used when
expanding macros inside the same crate. Tweak the name to avoid
warnings about unused variables.

* Update README example code

Wow, it's shameful to admit it, but the code does not even compile.
Make sure it's tested, fix it, and update README.

* Clean up thread safety section
* Update implementation safety comments
* Rename workflow for better badge

* Drop badges from Cargo.toml

Since we no longer use these CI platforms, and GitHub Actions are
not supported yet... Well, it's not a big deal if we leave them out.