Allow owned setters once again
This reverts commit 88909072b59031c5b37f72ac2243fa3a74b02225.

Partially revert the changes from that commit, bringing back setting
value through Borrow<T>. I have finally realized the background for
this feature. It allows efficient setting of value types, without
all these unnecessary references in initializer expressions.

Yeah, it does make it weird when "Copy" types get passed into set(),
but honestly it's just weird to expect that capturing behavior to
be used in the wild. Normally you don't combine set() and get() for
the same variable in one piece of code. If you do then why not just
use regular variables where static scoping works just fine?

Note that DynamicCell::set_guard() still accepts a reference to some
local variable, and DynamicCell::with_static() still initializes the
cell with a reference to static object. This is really important for
memory safety. We cannot make the guard own the value set because
the guard can be moved around and to allow that we'll need to pin
the set value somehow. "Pinning" it in static memory or on the stack
is the most efficient way.