Bump version to 1.0.0 (#11)
Actually, you know what? I'm pretty confident that this is the API
that I want to use, so let's be bold and go for stabilizing. It's
semver after all, I can always go and put out 2.0 if I find that
some APIs have to be broken. It's not like I'm going to piss off
"the community" by doing this, lol