Use `const fn` to create DynamicVariables (#19)
* Use "const fn" to create DynamicVariables

Instead of exposing a public field, use "const fn" for initialization.
This bump minimal rustc version to 1.31, but I don't particularly care
for anything but current stable, neither do rustc maintainers.

It's not a big deal since we already require 1.30+ because of "vis"
specifiers used in macros. And we probably need 1.31+ anyway because
we use 2018 edition.

* Conceal "cell" field of DynamicVariable

Now that we have a proper constructor, we don't need to expose this
field directly anymore.