Hide static initialization behind a "static-init" feature
I'm not fully ready to expose this feature in 1.0.0 since I'm still
in the process of figuring out the access API for variables that are
guaranteed to be initialized. I don't want it to have Option in it,
but it's not as easy to implement. So for now I'm going to hide this
capability behind a feature gate.