Make using static initialization a compiler error
Unfortunately, there is no nice way to inject cfg() check into expanded
macros. If you write #[cfg(...)] of cfg!(...) then those conditions end
up checked in the context of the user crate, checking *it's* features,
not features of fluid-let.

The only way around it is to provide a different definition of the
fluid_let! macro. It's not very pretty to maintain, but what choice
do I have here...