Gate static initialization behind `"static-init"` feature (#21)
* Hide static initialization behind a "static-init" feature

I'm not fully ready to expose this feature in 1.0.0 since I'm still
in the process of figuring out the access API for variables that are
guaranteed to be initialized. I don't want it to have Option in it,
but it's not as easy to implement. So for now I'm going to hide this
capability behind a feature gate.

* Make using static initialization a compiler error

Unfortunately, there is no nice way to inject cfg() check into expanded
macros. If you write #[cfg(...)] of cfg!(...) then those conditions end
up checked in the context of the user crate, checking *it's* features,
not features of fluid-let.

The only way around it is to provide a different definition of the
fluid_let! macro. It's not very pretty to maintain, but what choice
do I have here...

* Document "static-init" feature in module-level docs

* Update existing docs to mention "static-init" requirements

* Skip doctests when "static-init" is disabled

Well, they are still kinda compiled, but the macros don't get expanded
and thus don't fail compilation.

* Update logging example to use static initialization

It's very weird to use that function intermediary, just use static
initialization with the default log level. However, we need to add
some extra hacks for doctests to work with and without the feature.

* Use "static-init" on for docs.rs

Make sure that public docs will be built as if the feature is enabled.

* Don't compile testing dependencies separately

Since we have moved benchmarks out into a separate crate, test runs
are quick as they are and we can simply "cargo test" them without
having a separate compilation stage.

* Exercise tests with and without "static-init" feature

* Mention "static-init" feature in changelog